Added new home page for consumers/providers
After log in, it redirects to the mainSearch.html page for both
providers and consumers. we can create a new page for providers later
(and the routing is already set up for that) but for now it just
re-routes both consumers and providers to mainSearch.html, and routes
back to the dummy main page when you log out.