Web client patch
Fixed so the listings for that user are loaded after login, fixed the
logic for searching/no results. The sidebar is still an issue (i.e. as
the decrease the width of the window the sidebar links are no longer
clickable) and we may need to just use a navbar on the left instead. But
you guys would probably know better