Changed "watch post"
More convenient to store the watched listings within a user as opposed
to storing users in a listing. Added some functionality to the web app.
Still need to make it so the default button color for the watch post
button is used until the user pushes it. Still some issues with "apply"
commands being executed by multiple functions.